% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d_soainew.R
\name{d_CohortAdjust}
\alias{d_CohortAdjust}
\title{d_CohortAdjust a function to adjust the redistributed open age group deaths according to cohort size differences detected in earlier ages}
\usage{
d_CohortAdjust(Dnew, OA, ACVVmatrix, OAL, DOA, reproduce.matlab = TRUE)
}
\arguments{
\item{Dnew}{the interim data.frame, as passed in by \code{d_KannistoSplitYrSex()}. It has the same columns as the standard Deaths object, and draft death counts for the open age group as split into triangles upt to age 130.}

\item{OA}{the open age for this particular year/sex}

\item{OAL}{the open age Lexis shape, \code{"RR"} or \code{"VV"}. If the original data were \code{"TL"} or \code{"TU"}, these will have been handled earlier and converted appropriately.}

\item{DOA}{deaths in the open age group, as passed in by \code{d_KannistoSplitYrSex()}. Used for scaling.}

\item{VVmaster}{a matrix as passed in from \code{d_soainew()} via \code{d_KannistoSplitYrSex()}. This is a PC matrix (VV data in period-cohort format).}
}
\value{
Dnew Same as received, but with open age deaths adjusted for cohort size.
}
\description{
This is described in Appendix C, section 2. We take a very different programming strategy here than that employed in the legacy matlab scripts.
}
\details{
\code{PC2AC()} is also called in a sort of novel way. PC data (VV shape) are put into the AC shape for easy named reference.
}

