% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d_long.R
\name{d_long}
\alias{d_long}
\title{d_long adds 0s up to age 130 for years of deaths with no open age}
\usage{
d_long(Deaths)
}
\arguments{
\item{Deaths}{the standard internal Deaths data.frame, probably early in its processing, but not necessarily.}
}
\value{
Deaths the same deaths data.frame, but expanded to include ages up to 130 except for Year-Sex combos with pen ages, which are untouched.
}
\description{
Lexis triangles are created for the upper ages even if lower ages are RR or VV or something else. There may be overlap if the last age is not TL or TU, but it will be a 0 overlap, so it's innocuous. Intermediate gaps are also filled in. Year-Sex combos with open ages are left untouched.
}

