% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d_sNx1.R
\name{d_sNx1_inner}
\alias{d_sNx1_inner}
\title{a function that splits age groups for a single sex and year}
\usage{
d_sNx1_inner(DSexYr, MPVERSION = 5)
}
\arguments{
\item{DSexYr}{a chunk of the standard Deaths object, a single year and sex of data.}
}
\value{
The same data, with
}
\description{
This is the inner function that gets iterated inside \code{d_sNx1()}. Some matlab oddities are preserved and others are not, pending tests. This function can be fed any single-sex-year chunk of deaths data and is innocuous if there is nothing to be done. If there is RR data to be split, any pre-existing TL, TL, single-age RR or other Lexis shapes are preserved. For any RR data with age intervals greater than 1, we return the single age RR estimates. See MP for methods description, or check out the in-house spline function \code{d_sNx1_spline()}.
}

