% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d_soainew.R
\name{d_soainew}
\alias{d_soainew}
\title{d_s0ainew a function to redistribute deaths in the open age group over higher age triangles.}
\usage{
d_soainew(Deaths, reproduce.matlab = TRUE)
}
\arguments{
\item{Deaths}{the standard internal Deaths data.frame object. It MUST include only TL and TU entries, except the open age, which is ideally either RR or VV.}

\item{reproduce.matlab}{logical. Default = \code{TRUE}. Should the quirks of the matlab implementation be reproduced?}
}
\value{
Deaths the standard internal Deaths data.frame, with open age entries gone and replaced with death triangles up to age 130.
}
\description{
This function takes deaths in the open age group and redistributes them over higher ages according to the Kannisto survival function as fit to a synthetic pseudo-survival function composed of the backward cumulative sum of old-age deaths preceding the open age. The matlab implementation of this contained many aspects that were out-of-protocol, and these have been reproduced, optionally. One can also toggle this off and get Tim's interpretation of the MP. This function depends on several other functions, most notably \code{d_KannistoSplitYrSex()}, which does the whole process for a given year and sex. The present function is more of a wrappers, though some prep work is done here. Two other functions are important, \code{d_remfluct()} and \code{d_CohortAdjust()}, which outsource ugly aspects of the redistribution method. \code{d_remfluct()} contains some out-of-MP aspects that have not been re-interpreted, whereas \code{d_CohortAdjust()} will toggle between the matlab and Tim's proposal. Other minor dependent functions include \code{d_KannistoSurvMin()}, \code{d_KannistoSurv()} (stored inside the former), and \code{PC2AC()}. This function should be called AFTER deaths have been standardized to triangles, and 0s imputed where necessary below the open age group. The open age group itself can either be RR or VV. If the open age group is declared as TL or TU, these are overwritten to RR and VV, respectively, for the time being.
}

