% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dateYear}
\alias{dateYear}
\title{extract year element from date class}
\usage{
dateYear(Date)
}
\arguments{
\item{Date}{a vector of Date class.}
}
\value{
vector of years as integers
}
\description{
This makes for cleaner looking code in functions such as \code{p_movedata()}
}

