% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d_soainew.R
\name{KannistoSurvMin}
\alias{KannistoSurvMin}
\title{KannistoSurvMin and KannistoSurv two functions for fitting and predicting the Kannisto survival function}
\usage{
KannistoSurvMin(pars, P = SyntheticSurvival, oa = OA, N = N,
  x = Ages4Kannisto)
}
\arguments{
\item{pars}{or \code{ab} a vector or list of the starting values for the a and b parameters of the Kannisto function. Default value of \code{c(a = .1, b = .1)}.}

\item{P}{the synthetic survival function, as passed in from \code{d_KannistoSplitYrSex()}.}

\item{oa}{the open age, as passed in from \code{d_KannistoSplitYrSex()}.}

\item{N}{the number of data points (same as \code{length(SyntheticSurvival)}).}

\item{x}{a vector of the ages of the data points used in \code{SyntheticSurvival}.}
}
\value{
\code{KannistoSurvMin()} returns a residual to be minimized in \code{optim()} or similar. \code{KannistoSurv()} returns the predicted survival values (which may likely need to be rescaled later) for a given a,b and age vector.
}
\description{
These two functions are called by \code{d_KannistoSplitYrSex()}.
}

