% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/p_rescale.R
\name{p_agg}
\alias{p_agg}
\title{a function to aggregate population counts to N-year age groups}
\usage{
p_agg(Pop, N = 5)
}
\arguments{
\item{Pop}{standard HMD Population \code{data.frame}, in single ages.}

\item{N}{the N-year age group you want.}
}
\description{
This function is the opposite of what HMD usually does! It exists for the sake of rescaling HMD intercensal estimates to known 5-year age groups for the sake of US states. This is a peculiar data situation: we have our own intercensal estimates for Jan 1 in single year age groups, but we also have official intercensals in 5-year age groups for mid-year. We'd first need to move mid year to jan 1, but in order to do a good job of this, we first split to single ages. We don't want to just take these jan1 single ages as-is because we think the HMD \code{p_ic()} single-age distribution is better, as it uses info on both left and right, whereas \code{p_split()} only uses info from right... It's really a nerdy detail that only a demographer would care about. Anyway, in order to use the 'moved' official intercensals, we need to group back to 5-year age groups, and that's what this function is for. This function will NOT be exported, and it will be treated as an internal for \code{p_rescale()}.
}
\details{
Not that this function does not send things to the usual 'abridged' case. age 0 is not treated differently. The only N that make sense are 5 or 10, really. Not that official intercensals were for ages 0-4, etc, for US states, so this is consistent for this limited use.
}

