% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/p_ecm_findOmega.R
\name{p_ecm_findOmega}
\alias{p_ecm_findOmega}
\title{p_ecm_findOmega finds the omega neede in order to do the extinct cohort method}
\usage{
p_ecm_findOmega(Deaths, l = 5, threshold = 0.5)
}
\arguments{
\item{Deaths}{The standard Deaths data.frame, in its final unrounded state after all processing.}

\item{l}{how many recent cohorts will be checked to compute the averages that determine omega? Default = 5.}

\item{threshold}{how many deaths do we consider 'close to 0'? Per MP, we use 0.5, which is reasonable, since we average over 15 values.}
}
\value{
omega The age at which cohorts will be considered extinct
}
\description{
This function only takes Deaths (completed) as its input. The extinct cohort method requires the determination of an age by which cohorts are considered extinct. This function finds that age and returns it so that \code{p_ecm()} can do its job cleaner. Until checked more thoroughly, this could be off by one.
}
\details{

}

