% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/p_ic.R
\name{p_ic}
\alias{p_ic}
\title{generalized intercensal estimation}
\usage{
p_ic(Pop, Deaths, Births, reproduce.matlab = FALSE)
}
\arguments{
\item{Pop}{the standard \code{Pop} object, after age intervals and unknown ages have been dealt with.}

\item{Deaths}{the standard \code{Deaths} IDB object, after computations finalized.}

\item{Births}{the standard \code{Births} IDB object, after computations finalized.}

\item{reproduce.matlab}{logical. Default \code{FALSE}. Should we reproduce all aspects of the matlab code?}
}
\value{
Pop a standard population data.frame, in single ages, including all years in the intercensal period. \code{C1} and \code{C2} are only included in this object if they were January 1 estimates.
}
\description{
This function performs intercensal estimation for all intercensal periods containing at least one full calendar year. All age intervals must be single. Some checking is done to ensure upper ages are high enough. For C1 we must have a single-age estimate for population at age 79 at the lowest. C2 needs a single age estimate for at least age \code{79 + N}, where N is the number of years from C1 to C2, inclusive. If either census has open age groups that get in the way of this, we throw an error and stop the function. Open age groups must be dealt with prior to calling this function in any case. Likewise, age intervals greater than one must be split prior to calling this function.
}
\details{


The only Matlab oddity that is potentially affected by \code{reproduce.matlab} is the handling of dates. Matlab does not handle leap years. This function makes calls to two other LexisDB functions, such as \code{ypart()}, \code{d_addCohortColumn}, \code{p_addCohortColumn()} and \code{resortPops()}. Most work, and all MP formulas are found in the inner function, \code{p_ic_inner()}, which takes two clean censuses as its inputs.
}

