% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_ic.R
\name{p_ic_inner}
\alias{p_ic_inner}
\title{generalized intercensal estimation between two censuses only}
\usage{
p_ic_inner(C1, C2, Deaths, Births, reproduce.matlab = FALSE)
}
\arguments{
\item{C1}{the standard \code{Pop} object, selected only for rows containing the first (left-side) census.}

\item{C2}{the standard \code{Pop} object, selected only for rows containing the second (right-side) census.}

\item{Deaths}{the standard \code{Deaths} IDB object, after computations finalized.}

\item{Births}{the standard \code{Births} IDB object, after computations finalized.}

\item{reproduce.matlab}{logical. Default \code{FALSE}. Should we reproduce all aspects of the matlab code?}
}
\value{
Pop a standard population data.frame, in single ages, including all years in the intercensal period. \code{C1} and \code{C2} are only included in this object if they were January 1 estimates.
}
\description{
This function is called inside \code{p_ic()} on individual census pairs specified to be interpolated. We assume that all age intervals are 1, and we do light checking for upper ages. For C1 we must have a single-age estimate for population at age 79 at the lowest. C2 needs a single age estimate for at least age \code{79 + N}, where N is the number of years from C1 to C2, inclusive. If either census has open age groups that get in the way of this, we throw an error and stop the function. Open age groups must be dealt with prior to calling this function in any case. Likewise, age intervals greater than one must be split prior to calling this function.
}
\details{
The only matlab oddity that is potentially affected by \code{reproduce.matlab} is the handling of dates. Matlab does not handle leap years. This function makes calls to two other LexisDB functions, such as \code{ypart()}, \code{resortPops()}, and \code{p_addCohortColumn()}.
}

