% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_postcensal.R
\name{p_postcensal}
\alias{p_postcensal}
\title{a function for postcensal population estimation}
\usage{
p_postcensal(Pop, Deaths, Births, MPVERSION = 5, reproduce.matlab = FALSE)
}
\arguments{
\item{Pop}{the standard Pop object, ideally with open age split, e.g., after running \code{p_srecm()} on the data.}

\item{Deaths}{the standard Deaths object, after all processing is completed.}

\item{Births}{the standard Births object, only needs to be specified in \code{MPVERSION >= 7}.}

\item{MPVERSION}{numeric, default value is 5. This only toggles error distribution for newborn cohorts, as in description above.}

\item{reproduce.matlab}{logical, defaults to \code{FALSE}. This only affects date handling in the calculation of MP variable \code{f2}.}
}
\description{
This function is in principle simple, because we assume no error. This function should be called prior to \code{p_sr()} or \code{p_srecm()} in all cases. All age groups in \code{Pop} must be single (no open ages or groups). This function calls a few other LexisDB functions, such as \code{d_addCohortColumn()}, \code{p_addCohortColumn()}, \code{ypart()}, and \code{resortPops()}.
}

