% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_precensal.R
\name{p_precensal}
\alias{p_precensal}
\title{a function for precensal population estimation}
\usage{
p_precensal(Pop, Deaths, Births, MPVERSION = 5, reproduce.matlab = FALSE)
}
\arguments{
\item{Pop}{the standard Pop object, ideally with open age split, e.g., after running \code{p_srecm()} on the data.}

\item{Deaths}{the standard Deaths object, after all processing is completed.}

\item{Births}{the standard Births object, only needs to be specified in \code{MPVERSION >= 7}.}

\item{MPVERSION}{numeric, default value is 5. This only toggles error distribution for newborn cohorts, as in description above.}

\item{reproduce.matlab}{logical, defaults to \code{FALSE}. This only affects date handling in the calculation of MP variable \code{f2}.}
}
\description{
This function is in principle simple, because we assume no error, except possibly in the case of newborn cohorts. Whether or not we redistribute error over newborn cohorts in the precensal period is toggled with the \code{MPVERSION} argument. At the time of this writing, \code{MPVERSION >= 7} will distribute error over these cohorts, otherwise, we assume no error. This function should be called near the end of population estimation in the country script. All age groups in \code{Pop} must be single (no open ages or groups). So, typically, one would already have run \code{p_srecm()} or similar prior to calling this function. A little triangle of data in the top left lexis region may be left empty, and this can be filled by re-calling \code{p_srecm()}, for example. This function calls a few other LexisDB functions, such as \code{d_addCohortColumn()}, \code{p_addCohortColumn()}, \code{ypart()}, and \code{resortPops()}.
}

