% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_split.R
\name{p_split_inner}
\alias{p_split_inner}
\title{split age intervals based on a future census}
\usage{
p_split_inner(C1, C2, Deaths, Births, a = 80, reproduce.matlab = FALSE)
}
\arguments{
\item{C1}{the standard \code{Pop} object, selected only for rows containing the first (left-side) census.}

\item{C2}{the standard \code{Pop} object, selected only for rows containing the second (right-side) census.}

\item{Deaths}{the standard \code{Deaths} IDB object, after computations finalized.}

\item{Births}{the standard \code{Births} IDB object.}

\item{a}{the age where extinct cohort populations will take over. No need to split intervals above this age.}

\item{reproduce.matlab}{logical. Default \code{FALSE}. Should we reproduce all aspects of the matlab code?}
}
\value{
C2 in single ages
}
\description{
This function implements the MP verbal description of splitting population counts in age groups. Basically, we add deaths going back in time from census 2 (right-side), which will give an estimate for the single-age populations in the interval \code{[a,a+n]}. We then rescale these single age estimates to sum to the C1 population count in the interval.
}
\details{
The only matlab oddity that is potentially affected by \code{reproduce.matlab} is the handling of dates. Matlab does not handle leap years. This function makes calls to two other LexisDB functions, such as \code{ypart()} and \code{p_addCohortColumn()}.
}

