% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_unk.R
\name{p_unk}
\alias{p_unk}
\title{p_unk distribute popualtion counts of unknown ages.}
\usage{
p_unk(Pop)
}
\arguments{
\item{Pop}{the standard internal Population data.frame, probably late down the production chain.}
}
\value{
Pop rows containing UNK are removed, and these population counts have been distributed proportionally over other deaths of known age.
}

