% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PC2AP.R
\name{PC2AP}
\alias{PC2AP}
\title{\code{PC2AP} convert a period-cohort (VV) matrix to an age-period matrix, wherein entries still represent the VV shape.}
\usage{
PC2AP(PCmatrix, agemin = 80, agemax = 130, Lexis = 1)
}
\arguments{
\item{PCmatrix}{a matrix (period by cohort) of quantities in VV shape. May work for other shapes too, but you'll need to experiment with Lexis = 1 or 2.}

\item{Lexis}{either 1 or 2. Per HMD internal convention, 1 indicates lower triangles and 2 indicates upper triangles. For population counts, 1 indicates January 1st and 2 indicates December 31st. Default is 1 as the function was designed for a specific use.}
}
\value{
a \code{APmatrix} a matrix in age-period format, with correctly labeled dimensions. Note that if the original contents were the VV shape that age is the lower left corner, and that the shape spans two ages. Draw a picture if necessary.
}
\description{
\code{PC2AP} is an auxiliary function available to call by any of the HMD function family. Default lexis value is 1, so watch out- if you choose wrong, then column labels in the AP matrix output will be off by 1. If you're not sure, draw a diagram. If using this to reshape freshly crafted extinct-cohort-method population counts (VV-death-derived), then use Lexis = 1 to get the age right. This is not intuitive because TU marks the age of the VV shape in the HMD, so you'd think that Lexis = 2 should be used, but more generally we want this behavior for the Lexis argument.
}
\author{
Tim Riffe \email{triffe@demog.berkeley.edu}
}

