% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/readInputDB.R
\name{readInputDB}
\alias{readInputDB}
\title{read in the InputDB, in its entirety: XXXpop, XXXdeath, XXXtadj, XXXbirth, XXXmonthly}
\usage{
readInputDB(WORKING = "/data/commons/hmd/HMDWORK/DNK", XXX = NULL,
  log.file = "", InputDB.name = "InputDB", save.bin = FALSE,
  verbose = TRUE, strict = TRUE)
}
\arguments{
\item{WORKING}{the file path the country working folder. The last part of this path is typically the country abbreviation}

\item{XXX}{the standard HMD country abbreviation. If absent, \code{XXX} is inferred from the file path as its last segment. This is key for concatenating file names.}

\item{log.file}{log message are sent to the console by default (\code{""}). If you want these message written to a log file instead, specify a full path, including filename and extension.}

\item{InputDB.name}{the name of the InputDB folder, default \code{"InputDB"}. Must minimally contain files \code{XXXpop.txt}, \code{XXXdeath.txt}, \code{XXXbirth.txt}.}

\item{save.bin}{logical, should a copy of the function output be saved to the \code{XXX/Rbin/} directory?}

\item{verbose.}{logical. Default = \code{TRUE}. Should informative messages print to the console. If \code{log.file} is specified, most messages don't go to the console anyway.}
}
\value{
a \code{list} of all InputDB items as R \code{data.frames}
}
\description{
\code{readInputDB()} is the function that starts off the LexisDB programs by getting everything into R in memory, doing several checks, cleaning, etc. This should standardize all inputs for use downstream.
}
\details{
this function is essentially version independent. It will read \code{XXXmonthly} births (v6) if present, and return them in the list, but does not need these. Dame with \code{XXXtadj}, as not all countries have undergone a territorial adjustment,
}

