% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateLDB.R
\name{testAgeMatch}
\alias{testAgeMatch}
\title{an age checker}
\usage{
testAgeMatch(ages1, ages2, desc, tyear = NULL, tsex = NULL)
}
\arguments{
\item{ages1}{first set of ages (lower bounds)}

\item{ages2}{second set of ages (lower bounds)}

\item{desc}{optional, set of the fly to identify errors}

\item{tyear}{optional, set of the fly to identify errors}

\item{tsex}{optional, set of the fly to identify errors}
}
\description{
this is a utility used many times in \code{computeLDB()}
}

