% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateLDB.R
\name{writeLDB}
\alias{writeLDB}
\title{write out LexisDB}
\usage{
writeLDB(ldb.by.sex, LDBpath, COUNTRY = NULL, strict = TRUE)
}
\value{
function called for its side effects: some files written.
}
\description{
This is a beta function.
}

