% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{ypart}
\alias{ypart}
\title{ypart function to determine the proportion of a year passed as of a particular date}
\usage{
ypart(Year, Month, Day, reproduce.matlab = TRUE, detect.mid.year = FALSE,
  detect.start.end = TRUE)
}
\arguments{
\item{Year}{4-digit year (string or integer)}

\item{Month}{month digits (string or integer, 1 or 2 characters)}

\item{Day}{Day of month digits (string or integer, 1 or 2 characters)}

\item{reproduce.matlab}{logical. Default TRUE. Assume 365 days in a year.}

\item{detect.mid.year}{logical. if \code{TRUE}, June 30 or July 1 will always return .5.}

\item{detect.start.end}{logical. default \code{TRUE}. Should Jan 1 always be 0 and Dec 31 always be 1?}
}
\description{
The fraction returned by this is used e.g. for intercensal estimates. Function uses 'lubridate' package to handle dates elegantly.
}
\details{

}

