% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/p_ecm.R
\name{p_ecm}
\alias{p_ecm}
\title{p_ecm a simple extinct cohort method function, for modularity}
\usage{
p_ecm(Pop, Deaths, a = 80, reproduce.matlab = FALSE)
}
\arguments{
\item{Pop}{The standard internal Population data.frame, as}

\item{Deaths}{after all processing is done. Completed triangles.}

\item{a}{lowest age for which EC estimates should be given. Default 80.}

\item{reproduce.matlab}{logical. default is \code{FALSE}. This affects only the border cohort between EC and SR.}
}
\value{
Population with old ages either filled in or imputed using the extinct cohort method
}
\description{
This implementation seeks to follow the MP. Deaths must be finalized. Populations must be in single ages and all years. This function covers area B from Figure 6 'Methods used for population estimates'. The actual work is odone by \code{p_ecm_inner()}
}
\details{


Conceivably, a population could have a long series with low open ages in the early part, but decent population data in the recent part. If we want a value of \code{A} lower than 80 for only part of the series, this can be achieved by subsetting \code{Dsex}, and making two calls to the function, then \code{rbind()}ing back together.
}

